Refactoring enums and adding interceptor method for saving in appointment request service
