+++ HealthService/src/main/java/rs/uns/acs/ftn/HealthService/model/Ambulance.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
+import rs.uns.acs.ftn.HealthService.model.enumeration.AmbStatusEnum;
 
 @Setter
 @Getter
