+++ HealthService/src/main/java/rs/uns/acs/ftn/HealthService/model/Doctor.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
+import rs.uns.acs.ftn.HealthService.model.enumeration.RoleEnum;
 
 @Setter
 @Getter
