--- HealthService/src/main/java/rs/uns/acs/ftn/HealthService/model/RoleEnum.java
@@ -1,5 +1,0 @@
-package rs.uns.acs.ftn.HealthService.model;
-
-public enum RoleEnum {
-    ROLE_NURSE, ROLE_DOCTOR
-}
