+++ HealthService/src/main/java/rs/uns/acs/ftn/HealthService/model/enumeration/AmbStatusEnum.java
@@ -1,4 +1,4 @@
-package rs.uns.acs.ftn.HealthService.model;
+package rs.uns.acs.ftn.HealthService.model.enumeration;
 public enum AmbStatusEnum {
     AVAILABLE, BUSY, PREPARING
 }
