+++ HealthService/src/main/java/rs/uns/acs/ftn/HealthService/model/enumeration/AppointmentRequestStatusEnum.java
@@ -1,0 +1,5 @@
+package rs.uns.acs.ftn.HealthService.model.enumeration;
+
+public enum AppointmentRequestStatusEnum {
+    PENDING, ACCEPTED, REJECTED
+}
