+++ HealthService/src/main/java/rs/uns/acs/ftn/HealthService/model/enumeration/EmergencyStatusEnum.java
@@ -1,4 +1,4 @@
-package rs.uns.acs.ftn.HealthService.model;
+package rs.uns.acs.ftn.HealthService.model.enumeration;
 
 public enum EmergencyStatusEnum {
     PENDING, REJECTED, ACCEPTED
