+++ HealthService/src/main/java/rs/uns/acs/ftn/HealthService/model/enumeration/RoleEnum.java
@@ -1,0 +1,5 @@
+package rs.uns.acs.ftn.HealthService.model.enumeration;
+
+public enum RoleEnum {
+    ROLE_NURSE, ROLE_DOCTOR
+}
